LayoutTests/fast/events/iframe-object-onload.html

<div> 
This test checks that onload events fire on iframe and object
elements when their contents finish loading, and that these are
separate event objects from those that fire on the body inside the
frame. If it passes, you will see 6 blocks below, each with a
three-line message that ends in false.
</div>
<div id="console">
</div>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

function log(s)
{
    document.getElementById("console").innerHTML += s + "<br>";
}

var innerEvent = nil;

function logOuter(msg, event)
{
    log(msg);
    log('Outer event same as inner: ' + (innerEvent == event));
    log('------');
}

function logInner(tag, event)
{
    parent.log("Fired load event on body inside " + tag); 
    parent.innerEvent = event;
}

</script>

<iframe 
    src="data:text/html,<html><body onload='parent.logInner(%22iframe%22, event)'>Inner</body></html>" 
    onload="logOuter('Fired event on iframe', event)">
</iframe>

<iframe 
    src="data:text/html,<body onload='parent.logInner(%22iframe%22, event); event.preventDefault()'>preventDefault</body>" 
    onload="logOuter('Fired event on iframe even though default was prevented', event)">
</iframe>

<iframe
    src="data:text/html,<body onload='parent.logInner(%22object%22, event); event.stopPropagation()'>stopPropagation</body>" 
    onload="logOuter('Fired event on object even though propagation was stopped', event)">
</iframe>

<object 
    type="text/html"
    data="data:text/html,<html><body onload='parent.logInner(%22object%22, event)'>Inner</body></html>" 
    onload="logOuter('Fired event on object', event)">
</object>

<object 
    type="text/html"
    data="data:text/html,<body onload='parent.logInner(%22object%22, event); event.preventDefault()'>preventDefault</body>" 
    onload="logOuter('Fired event on object even though default was prevented', event)">
</object>

<object 
    type="text/html"
    data="data:text/html,<body onload='parent.logInner(%22object%22, event); event.stopPropagation()'>stopPropagation</body>" 
    onload="logOuter('Fired event on object even though propagation was stopped', event)">
</object>
